+++ src/spine/controller/implementations/pid/pid_tuner.hpp
@@ -38,16 +38,16 @@ public:
     /// outputRange: min and max values of the output that can be used to control the system (0, 255 for analogWrite)
     /// znMode: Ziegler-Nichols tuning mode (znModeBasicPID, znModeLessOvershoot, znModeNoOvershoot)
     /// tuningCycles: number of cycles that the tuning runs for (optional, default is 10)
-    void set_target_input_value(double target) { targetInputValue = target; }
+    void set_target_input_value(float target) { targetInputValue = target; }
     void set_loop_interval(k_time_ms interval) { loopInterval = interval; }
 
     /// Set hysteresis; the minimal overshoot over `targetInputValue`
     /// This is to discourage a sensor bouncing below and over the `targetInputValue` causing a cycle with a false
     /// length
-    void set_hysteresis(double hysteresis) { targetHysteresis = std::fabs(hysteresis); }
+    void set_hysteresis(float hysteresis) { targetHysteresis = std::fabs(hysteresis); }
 
     /// Set output range
-    void setOutputRange(double lower_limit, double upper_limit) {
+    void setOutputRange(float lower_limit, float upper_limit) {
         minOutput = lower_limit;
         maxOutput = upper_limit;
     }
@@ -62,12 +62,12 @@ public:
     /// Automatically tune PID
     /// This function must be run in a loop at the same speed as the PID loop being tuned
     /// See README for more details - https://github.com/jackw01/arduino-pid-autotuner/blob/master/README.md
-    double do_autotune(double input, k_time_ms ms);
+    float do_autotune(float input, k_time_ms ms);
 
     /// Get results of most recent tuning
-    double get_Kp() const { return kp; };
-    double get_Ki() const { return ki; };
-    double get_Kd() const { return kd; };
+    float get_Kp() const { return kp; };
+    float get_Ki() const { return ki; };
+    float get_Kd() const { return kd; };
 
     /// Is the tuning loop finished?
     bool is_finished() const { return (cycle >= cycles); }
@@ -77,20 +77,20 @@ public:
 
 private:
 private:
-    double targetInputValue = 0;
-    double targetHysteresis = 0;
+    float targetInputValue = 0;
+    float targetHysteresis = 0;
     k_time_ms loopInterval = {};
-    double minOutput = 0, maxOutput = 0;
+    float minOutput = 0, maxOutput = 0;
     Aggressiveness znMode = Aggressiveness::NoOvershoot;
     int cycles = 10;
 
     int cycle = 0; // current cycle
     bool output = false; // is output on
-    double outputValue = 0; // current output value
+    float outputValue = 0; // current output value
     k_time_ms t1 = {}, t2 = {}, tHigh = {}, tLow = {};
-    double max = 0, min = 0;
-    double pAverage = 0, iAverage = 0, dAverage = 0;
-    double kp = 0, ki = 0, kd = 0;
+    float max = 0, min = 0;
+    float pAverage = 0, iAverage = 0, dAverage = 0;
+    float kp = 0, ki = 0, kd = 0;
 };
 
 } // namespace spn::controller
